Revert "Revert "Add Windows support for the Python API (#1811)" (#1959)"
This reverts commit e231201cfc9638c38dbf50b02973b764070e4521.

+++ apis/python/src/tiledbsoma/io/ingest.py
@@ -1163,7 +1163,7 @@ def _write_dataframe(
         else:
             df.rename(columns={"index": id_column_name}, inplace=True)
 
-    df[SOMA_JOINID] = np.asarray(axis_mapping.data)
+    df[SOMA_JOINID] = np.asarray(axis_mapping.data, dtype=np.int64)
 
     df.set_index(SOMA_JOINID, inplace=True)
